feat: Extract search components and utils to ducks/search
