+++ src/ducks/apps/components/Sections/Sections.jsx
@@ -4,120 +4,15 @@ import Fuse from 'fuse.js'
 import sortBy from 'lodash/sortBy'
 import debounce from 'lodash/debounce'
 
-import Input from 'cozy-ui/transpiled/react/Input'
-import InputGroup from 'cozy-ui/transpiled/react/InputGroup'
-import Label from 'cozy-ui/transpiled/react/Label'
 import Icon from 'cozy-ui/transpiled/react/Icon'
 import { translate, useI18n } from 'cozy-ui/transpiled/react/I18n'
-import useBreakpoints from 'cozy-ui/transpiled/react/hooks/useBreakpoints'
 import AppSections from 'cozy-ui/transpiled/react/AppSections'
 import * as filterUtils from 'cozy-ui/transpiled/react/AppSections/search'
 
 import flag from 'cozy-flags'
-import StoreAppItem from './components/StoreAppItem'
 
-const fillIndices = (str, indices) => {
-  let cur = 0
-  let res = []
-  for (let idx of indices) {
-    if (idx[0] != cur) {
-      res.push({ idx: [cur, idx[0]], mark: false })
-    }
-    res.push({ idx, mark: true })
-    cur = idx[1]
-  }
-  if (cur != str.length) {
-    res.push({ idx: [cur, str.length], mark: false })
-  }
-
-  return res
-}
-
-const dumpmatches = result => {
-  for (let m of result.matches) {
-    const allindices = fillIndices(m.value, m.indices)
-    // eslint-disable-next-line no-console
-    console.log(
-      allindices
-        .map(o => {
-          const [start, end] = o.idx
-          const substr = m.value.substring(start, end)
-          return o.mark ? `**${substr}**` : substr
-        })
-        .join('')
-    )
-  }
-}
-
-const SearchField = ({ onChange, value }) => {
-  const { t } = useI18n()
-  const { isMobile } = useBreakpoints()
-  const handleChange = useCallback(
-    ev => {
-      onChange(ev.target.value)
-    },
-    [onChange]
-  )
-  return (
-    <>
-      {!isMobile ? (
-        <Label className="u-di u-mr-half" htmlFor="discover-search">
-          {t('discover-search-field.label')}
-        </Label>
-      ) : null}
-      <InputGroup
-        className={isMobile ? '' : 'u-mb-1'}
-        prepend={
-          isMobile ? (
-            <Icon icon="magnifier" className="u-pl-1 u-coolGrey" />
-          ) : null
-        }
-      >
-        <Input
-          id="discover-search"
-          placeholder={t('discover-search-field.placeholder')}
-          onChange={handleChange}
-          type="text"
-          value={value}
-        />
-      </InputGroup>
-    </>
-  )
-}
-
-const SearchResults = ({ searchResults, onAppClick }) => {
-  const sortedSortResults = useMemo(() => {
-    return sortBy(searchResults, result => result.score)
-  }, [searchResults])
-  return (
-    <div className="u-mv-1 u-flex u-flex-wrap">
-      {sortedSortResults.map(result => {
-        const app = result.item
-        return flag('store.show-search-score') ? (
-          <div>
-            <StoreAppItem
-              onClick={() => onAppClick(app.slug)}
-              key={app.slug}
-              app={app}
-            />
-            <small
-              onClick={() => dumpmatches(result)}
-              title={JSON.stringify(result.matches, null, 2)}
-            >
-              {result.score.toFixed(2)}
-            </small>
-          </div>
-        ) : (
-          <StoreAppItem
-            onClick={() => onAppClick(app.slug)}
-            key={app.slug}
-            app={app}
-          />
-        )
-      })}
-    </div>
-  )
-}
+import { dumpMatches } from 'ducks/search/utils'
+import { SearchField, SearchResults } from 'ducks/search/components'
 
 /**
  * Shows a list of apps grouped by categories.
@@ -158,7 +53,6 @@ export class Sections extends Component {
       includeScore: true,
       includeMatches: true,
       minMatchCharLength: 3,
-      distance: 10,
       threshold: 0.2,
       ignoreLocation: true,
       keys: [
