+++ src/ducks/apps/components/StoreAppItem.jsx
@@ -2,7 +2,7 @@ import React from 'react'
 import { useI18n } from 'cozy-ui/transpiled/react/I18n'
 import AppTile from 'cozy-ui/transpiled/react/AppTile'
 
-import { getTranslatedManifestProperty } from '../helpers'
+import { getTranslatedManifestProperty } from './helpers'
 
 const StoreAppItem = ({ app, onClick }) => {
   const { t } = useI18n()
