+++ src/ducks/apps/components/helpers.js
