+++ src/ducks/apps/components/helpers.spec.js
